common generator
