+++ common/pluginsInfo.js
